move supabase client file
