+++ src/helpers/index.ts
@@ -6,5 +6,4 @@ export * from './getTvUtils';
 export * from './showFilterUtils';
 export * from './showSortUtils';
 export * from './stringFormatUtils';
-export * from './supabaseClient';
 export * from './validationUtils';
